corrections and improvement cookPerfil
correction and improvement of cook Profile.html, customization of errors 500, 403, 404, 400, modal to view recipes

+++ ProyectoIntegradoFlavorSync/src/main/java/com/example/demo/model/recipeModel.java
@@ -231,7 +231,7 @@ public class recipeModel {
 		this.category = category;
 	}
 
-	public double getAverageRating() {
+	public float getAverageRating() {
 		return averageRating;
 	}
